Added attendance list
