+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Cal</title>
     <link rel="stylesheet" href="styles.css">
-    <script src="./main.js" defer></script>
+    <script type="module" src="./js/main.js" defer></script>
 </head>
 
 <body>
@@ -41,7 +41,7 @@
             <td class="lec" colspan="2">PCC Lab</td>
             <td class="lec">Maths</td>
             <td class="lec">Phy</td>
-            <td rowspan="5">RECESS</td>
+            <td class="lec" rowspan="5">RECESS</td>
             <td class="lec" colspan="2">EG Lab</td>
             <td class="lec">EG</td>
             <td></td>
@@ -84,10 +84,9 @@
             <td class="lec">Phy</td>
             <td class="lec" colspan="2">PP Lab</td>
         </tr>
-
     </table>
-    <div class="container"></div>
-    <div class="attendanceContainer"></div>
+    <div class="promptContainer"></div>
+    <div class="attContainer"></div>
 </body>
 
 </html>
